@@ -13,7 +13,7 @@ class ExportsController < ChouetteController
       index! do 
         build_breadcrumb :index
       end
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
@@ -25,7 +25,7 @@ class ExportsController < ChouetteController
       show! do 
         build_breadcrumb :show
       end
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
@@ -35,7 +35,7 @@ class ExportsController < ChouetteController
   def destroy    
     begin
       destroy!
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
@@ -49,7 +49,7 @@ class ExportsController < ChouetteController
     OpenURI::Buffer.const_set 'StringMax', 0
     begin
       send_file open(resource.file_path), { :type => "application/#{resource.filename_extension}", :disposition => "attachment", :filename => resource.filename }
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
@@ -67,7 +67,7 @@ class ExportsController < ChouetteController
       @compliance_check = resource
       build_breadcrumb :compliance_check
       render "compliance_checks/report"
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
