@@ -10,7 +10,7 @@ class ImportTasksController < ChouetteController
     @available_imports = available_imports
     begin
       new!
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t('iev.exception.default')
       redirect_to referential_path(@referential)
@@ -23,7 +23,7 @@ class ImportTasksController < ChouetteController
       create! do |success, failure|
         success.html { redirect_to referential_imports_path(@referential) }
       end
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t('iev.exception.default')
       redirect_to referential_path(@referential)
