@@ -14,7 +14,7 @@ class ImportsController < ChouetteController
       index! do
         build_breadcrumb :index
       end
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
@@ -26,7 +26,7 @@ class ImportsController < ChouetteController
       show! do
         build_breadcrumb :show
       end
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
@@ -36,7 +36,7 @@ class ImportsController < ChouetteController
   def destroy
     begin
       destroy!
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
@@ -50,7 +50,7 @@ class ImportsController < ChouetteController
     OpenURI::Buffer.const_set 'StringMax', 0
     begin
       send_file open(resource.file_path), { :type => "application/#{resource.filename_extension}", :disposition => "attachment", :filename => resource.filename }
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
@@ -62,7 +62,7 @@ class ImportsController < ChouetteController
       @rule_parameter_set = resource.rule_parameter_set
       build_breadcrumb :rule_parameter_set
       render "rule_parameter_sets/show"
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
@@ -80,7 +80,7 @@ class ImportsController < ChouetteController
       @compliance_check = resource
       build_breadcrumb :compliance_check
       render "compliance_checks/report"
-    rescue Ievkit::Error, Faraday::Error => error
+    rescue Ievkitdeprecated::Error, Faraday::Error => error
       logger.error("Iev failure : #{error.message}")
       flash[:error] = t(error.locale_for_error) if error.methods.include? :locale_for_error
       redirect_to referential_path(@referential)
